+++ R/Learner.R
@@ -551,6 +551,8 @@ Learner = R6Class("Learner",
     #' Controls how to execute the code in internal train and predict methods.
     #' Must be a named character vector with names `"train"` and `"predict"`.
     #' Possible values are `"none"`, `"try"`, `"evaluate"` (requires package \CRANpkg{evaluate}) and `"callr"` (requires package \CRANpkg{callr}).
+    #' When encapsulation is activated, a fallback learner must be set.
+    #' If no learner is set in `$fallback`, the default fallback learner is used (see `mlr_reflections$task_types`).
     #' See [mlr3misc::encapsulate()] for more details.
     encapsulate = function(rhs) {
       default = c(train = "none", predict = "none")
@@ -562,6 +564,11 @@ Learner = R6Class("Learner",
       assert_character(rhs)
       assert_names(names(rhs), subset.of = c("train", "predict"))
       private$.encapsulate = insert_named(default, rhs)
+
+      if (is.null(private$.fallback)) {
+        task_type = self$task_type
+        self$fallback = lrn(mlr_reflections$task_types[task_type, fallback], predict_type = self$predict_type)
+      }
     },
 
     #' @field fallback ([Learner])\cr