+++ R/mlr_reflections.R
@@ -13,7 +13,7 @@
 #' * `task_types` (`data.table()`)\cr
 #'   Table with task type (`"type"`), the implementing package (`"pkg"`), and the names of the generators
 #'   of the corresponding [Task] (`"task"`), [Learner] (`"learner"`),
-#'   [Prediction] (`"prediction"`) and [Measure] (`"measure"`).
+#'   [Prediction] (`"prediction"`), [Measure] (`"measure"`) and fallback [Learner].
 #'
 #' * `task_feature_types` (named `character()`)\cr
 #'   Vector of base R types supported as [Task] features, named with a 3 letter abbreviation.
@@ -78,10 +78,10 @@ local({
   ### Task
   # task types + constructors
   mlr_reflections$task_types = rowwise_table(.key = "type",
-    ~type,          ~package, ~task,              ~learner,         ~prediction,          ~prediction_data,         ~measure,
-    "regr",         "mlr3",   "TaskRegr",         "LearnerRegr",    "PredictionRegr",     "PredictionDataRegr",     "MeasureRegr",
-    "classif",      "mlr3",   "TaskClassif",      "LearnerClassif", "PredictionClassif",  "PredictionDataClassif",  "MeasureClassif",
-    "unsupervised", "mlr3",   "TaskUnsupervised", "Learner",        NA_character_,        NA_character_,            NA_character_
+    ~type,          ~package, ~task,              ~learner,         ~prediction,          ~prediction_data,         ~measure,           ~fallback,
+    "regr",         "mlr3",   "TaskRegr",         "LearnerRegr",    "PredictionRegr",     "PredictionDataRegr",     "MeasureRegr",      "regr.featureless",
+    "classif",      "mlr3",   "TaskClassif",      "LearnerClassif", "PredictionClassif",  "PredictionDataClassif",  "MeasureClassif",   "classif.featureless",
+    "unsupervised", "mlr3",   "TaskUnsupervised", "Learner",        NA_character_,        NA_character_,            NA_character_,       NA_character_
   )
 
   mlr_reflections$task_feature_types = c(